requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
anthropic>=0.18.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jinja2>=3.1.0
pillow>=10.0.0
youtube-transcript-api>=0.6.0
pandas>=2.0.0
langdetect>=1.0.9
deep-translator>=1.11.0
playwright>=1.40.0
aiohttp>=3.9.0
